upadte handle bars and login

+++ public/js/deletesong.js
@@ -1,17 +1,19 @@
 var savedBtn = document.querySelector("#song");
-savedBtn.addEventListener("click", function () {
-    if(savedBtn.textContent === "❤") {
-        savedBtn.textContent = "♡"
-      const artist_id = savedBtn.getAttribute('data-id');
-      fetch(`/api/favorites/${artist_id}`, {
-        method: 'DELETE'
-    })
-    .then(response=>{
-        console.log(response)
-        if(response.ok){
-            document.location.replace('/playlist');
+if (savedBtn) {
+    savedBtn.addEventListener("click", function () {
+        if (savedBtn.textContent === "❤") {
+            savedBtn.textContent = "♡"
+            const artist_id = savedBtn.getAttribute('data-id');
+            fetch(`/api/favorites/${artist_id}`, {
+                method: 'DELETE'
+            })
+                .then(response => {
+                    console.log(response)
+                    if (response.ok) {
+                        document.location.replace('/playlist');
+                    }
+
+                });
         }
-       
     });
-    }
-})
+}
